requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Django==4.0.6
idna==3.3
Pillow==9.2.0
pip==22.2
setuptools==63.2.0
sqlparse==0.4.2
urllib3==1.26.10
wheel==0.37.1
